% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Melasto_probe_set-package.R
\docType{package}
\name{Clean_Melasto_probe_set-package}
\alias{Clean_Melasto_probe_set}
\alias{Clean_Melasto_probe_set-package}
\title{Clean_Melasto_probe_set: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LPDagallier/Clean_Melasto_probe_set}
  \item Report bugs at \url{https://github.com/LPDagallier/Clean_Melasto_probe_set/issues}
}

}
\author{
\strong{Maintainer}: Leo-Paul Dagallier \email{leopauldagallier@gmail.com} (\href{https://orcid.org/0000-0002-3270-1544}{ORCID}) [copyright holder]

}
\keyword{internal}
